+++ app/services/contact/set_zazo_info_by_owner.rb
@@ -1,0 +1,11 @@
+class Contact::SetZazoInfoByOwner
+  attr_reader :owner
+
+  def initialize(owner)
+    @owner = Owner.new(owner)
+  end
+
+  def do
+    owner.contacts.map { |contact| Contact::SetZazoInfoByContact.new(contact).do }
+  end
+end