+++ app/workers/update_new_contacts_by_owner_worker.rb
@@ -5,7 +5,7 @@ class UpdateNewContactsByOwnerWorker
 
   def self.perform(owner_mkey)
     WriteLog.info(self, "resque job was executed for owner_mkey=#{owner_mkey}")
-    Contact::SetZazoIdAndMkeyByOwnerContacts.new(owner_mkey).do
+    Contact::SetZazoInfoByOwner.new(owner_mkey).do
     Score::CalculationByOwner.new(owner_mkey).do
   end
 end